Create CreateUserDto for User entity
CreateUserDto, representing the data required for creating a new user. The class is annotated with Lombok's @Value and @Builder to generate the necessary methods and builders. The fields include name, birthday, email, password, role, and gender, aligning with the corresponding attributes of the User entity. This DTO will facilitate the transfer of user creation data between different layers of the application, such as controllers and services.