Implement CreateUserValidator for user creation validation
Implementing the Validator<CreateUserDto> interface, to handle validation logic specific to the creation of user data. The class includes a private constructor, a static instance, and an isValid method. In the isValid method, gender validation is performed, and an error is added to the ValidationResult if the gender is invalid. The getInstance method provides a singleton instance of the validator. This commit establishes the groundwork for custom validation of CreateUserDto objects during user creation.