Enhance Gender enum with find method
Extend the enumeration by gender by adding a static find method that allows you to extract the enumeration by gender based on its name. The find method uses Java Streams and Optional, which provides a more flexible and readable way to search an enumeration by gender by name. This is done because the Gender.valueOf(String name) method can return an exception. So we are working with null safety